add sanity check before running remove command

+++ scripts/update.sh
@@ -1,7 +1,8 @@
 #!/usr/bin/env bash
 # clean
-rm -rf $HOME/conv
-rm -rf $HOME/stats
+if [ -d "$HOME/conv" ]; then
+    rm -rf $HOME/conv
+fi
 
 # clone new code, start the system and stop
 git clone https://github.com/parallel-ml/conv.git $HOME/conv
